Use webgpu-utils to upload game config to GPU

+++ shaders/config.wgsl
@@ -10,7 +10,7 @@ struct Config {
     MINERAL_LEVEL_HEIGHT: i32,
     RELATIVE_THRESHOLD:   u32,
     REPRODUCTION_COST:    i32,
-    MUTATION_RATE:        u32,
+    MUTATION_RATE:        f32,
 }
 
 fn isValidPos(pos: vec2i) -> bool {
